# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "CadastralParcelResponse": {
        "properties": {
          "cadastral_id": {
            "title": "Cadastral Id",
            "type": "string"
          },
          "commune_code": {
            "title": "Commune Code",
            "type": "string"
          },
          "department_code": {
            "title": "Department Code",
            "type": "string"
          },
          "geometry": {
            "additionalProperties": true,
            "title": "Geometry",
            "type": "object"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ParcelProvenanceResponse"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "stated_area_m2": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stated Area M2"
          }
        },
        "required": [
          "cadastral_id",
          "department_code",
          "commune_code",
          "prefix",
          "section",
          "number",
          "stated_area_m2",
          "geometry",
          "provenance"
        ],
        "title": "CadastralParcelResponse",
        "type": "object"
      },
...
      "EntityDetailResponse": {
        "properties": {
          "area_m2": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area M2"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "type": "array"
          },
          "center": {
            "items": {
              "type": "number"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Center",
            "type": "array"
          },
          "commune_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune Code"
          },
          "commune_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune Name"
          },
          "department_code": {
            "title": "Department Code",
            "type": "string"
          },
          "entity_type": {
            "enum": [
              "parcel",
              "building",
              "property_unit"
            ],
            "title": "Entity Type",
            "type": "string"
          },
          "geometry": {
            "additionalProperties": true,
            "title": "Geometry",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "related_entities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Related Entities",
            "type": "array"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "id",
          "entity_type",
          "label",
          "commune_code",
          "commune_name",
          "department_code",
          "center",
          "bbox",
          "geometry",
          "properties",
          "related_entities",
          "sources"
        ],
        "title": "EntityDetailResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParcelProvenanceResponse": {
        "properties": {
          "data_source_id": {
            "title": "Data Source Id",
            "type": "string"
          },
          "layer": {
            "title": "Layer",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "raw_asset_id": {
            "title": "Raw Asset Id",
            "type": "integer"
          },
          "record_checksum": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Record Checksum",
            "type": "string"
          },
          "release_id": {
            "title": "Release Id",
            "type": "string"
          },
          "release_key": {
            "title": "Release Key",
            "type": "string"
          },
          "repair_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repair Method"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "source_published_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Published On"
          },
          "source_row_number": {
            "title": "Source Row Number",
            "type": "integer"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "transformation_version": {
            "title": "Transformation Version",
            "type": "string"
          },
          "was_repaired": {
            "title": "Was Repaired",
            "type": "boolean"
          }
        },
        "required": [
          "data_source_id",
          "release_id",
          "release_key",
          "source_published_on",
          "raw_asset_id",
          "layer",
          "source_url",
          "object_key",
          "sha256",
          "source_row_number",
          "record_checksum",
          "transformation_version",
          "was_repaired",
          "repair_method"
        ],
        "title": "ParcelProvenanceResponse",
        "type": "object"
      },
...
      "PublicationRequest": {
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "PublicationRequest",
        "type": "object"
      },
...
      "RollbackRequest": {
        "properties": {
          "bundle_id": {
            "pattern": "^brittany:[0-9a-f-]{36}$",
            "title": "Bundle Id",
            "type": "string"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "bundle_id"
        ],
        "title": "RollbackRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Immo Opportunities API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/brittany/publications": {
      "post": {
        "operationId": "create_brittany_publication_api_v1_admin_brittany_publications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Brittany Publication Api V1 Admin Brittany Publications Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Brittany Publication",
        "tags": [
          "brittany-pilot"
        ]
      }
    },
...
    "/api/v1/admin/brittany/rollback": {
      "post": {
        "operationId": "rollback_brittany_publication_api_v1_admin_brittany_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Rollback Brittany Publication Api V1 Admin Brittany Rollback Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rollback Brittany Publication",
        "tags": [
          "brittany-pilot"
        ]
      }
    },
    "/api/v1/admin/brittany/withdraw": {
      "post": {
        "operationId": "withdraw_brittany_publication_api_v1_admin_brittany_withdraw_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Withdraw Brittany Publication Api V1 Admin Brittany Withdraw Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw Brittany Publication",
        "tags": [
          "brittany-pilot"
        ]
      }
    },
...
    "/api/v1/cadastre/parcels/{cadastral_id}": {
      "get": {
        "operationId": "cadastral_parcel_api_v1_cadastre_parcels__cadastral_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cadastral_id",
            "required": true,
            "schema": {
              "title": "Cadastral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CadastralParcelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cadastral Parcel",
        "tags": [
          "cadastre"
        ]
      }
    },
...
    "/api/v1/parcels/{parcel_id}": {
      "get": {
        "operationId": "parcel_api_v1_parcels__parcel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "parcel_id",
            "required": true,
            "schema": {
              "title": "Parcel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parcel",
        "tags": [
          "real-map"
        ]
      }
    },
...
  }
}